backend/requirements.txt
flask
flask-cors
pdfplumber
spacy
nltk
python-dotenv
numpy
pandas
scikit-learn
faiss-cpu
sentence-transformers
langchain
langchain-openai
langchain-community
regex
werkzeug